example_project/requirements.txt
# Requirements specifically for the example app
Django>=1.11
django-debug-toolbar
django-polymorphic>=2.0
djangorestframework
inflection
pluggy
py
pyparsing
pytz
sqlparse
django-filter>=2.0
dj-database-url==0.5.0
factory-boy==2.12.0
